# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server TODO server.",
        "title": "TODO API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/labels": {
            "get": {
                "description": "Get all labels.",
                "tags": [
                    "labels"
                ],
                "summary": "Get labels.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.getLabelResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create label.",
                "tags": [
                    "labels"
                ],
                "summary": "Create label.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Label",
                        "name": "label",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createLabelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.getLabelResponse"
                        }
                    }
                }
            }
        },
        "/labels/{label_id}": {
            "get": {
                "description": "Get label.",
                "tags": [
                    "labels"
                ],
                "summary": "Get label.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Label ID",
                        "name": "label_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.getLabelResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update label.",
                "tags": [
                    "labels"
                ],
                "summary": "Update label.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Label ID",
                        "name": "label_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Label",
                        "name": "label",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createLabelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.getLabelResponse"
                        }
                    }
                }
            }
        },
        "/labels/{label_id}/tasks": {
            "get": {
                "description": "Get tasks for label.",
                "tags": [
                    "labels"
                ],
                "summary": "Get tasks for label.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Label ID",
                        "name": "label_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.getTaskResponse"
                            }
                        }
                    }
                }
            }
        },
...
        "/tasks/{task_id}/comments": {
            "get": {
                "description": "Get comments for task by Task ID.",
                "tags": [
                    "comments"
                ],
                "summary": "Get task comments.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.getCommentResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create comment for task by Task ID.",
                "tags": [
                    "comments"
                ],
                "summary": "Create task comment.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createCommentRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/tasks/{task_id}/comments/{comment_id}": {
            "get": {
                "description": "Get task comment by Task ID and Comment ID.",
                "tags": [
                    "comments"
                ],
                "summary": "Get task comment.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.getCommentResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update task comment by Task ID and Comment ID.",
                "tags": [
                    "comments"
                ],
                "summary": "Update comment.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createCommentRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Delete task comment by Task ID and Comment ID.",
                "tags": [
                    "comments"
                ],
                "summary": "Delete comment from task.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "server.createCommentRequest": {
            "type": "object",
            "required": [
                "description",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "server.createLabelRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
...
        "server.getCommentResponse": {
            "type": "object",
            "properties": {
                "comment_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "task_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "server.getLabelResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "label_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "server.getTaskResponse": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "task_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}